Make the generated CRUD code more RESTful
Currently, the generated CRUD code uses just GET and POST and completely ignores the other HTTP methods. Furthermore, the generated URLs do not comply to the common practice of having a "collection" URL a number of entity URLs. This could be changed easily by leveraging Symfony's handling of the "_method"  parameter which allows to override the HTTP method so that also standard browsers can emulate PUT and DELETE requests.